--- src/ui/src/app/components/lint-rule-labels.ts
@@ -1,14 +1,0 @@
-import type { GraphVisualizationProjectConfigurationLintRuleEntry } from "../../graph/index.js";
-
-/**
- * Returns the canonical badge label for lint rule fixability.
- */
-export function getLintFixableBadgeLabel(
-    fixable: GraphVisualizationProjectConfigurationLintRuleEntry["fixable"]
-): string | null {
-    if (fixable === null) {
-        return null;
-    }
-
-    return "fixable";
-}
