+++ src/ui/test/lint-rule-levels.test.ts
@@ -2,6 +2,7 @@ import assert from "node:assert/strict";
 import test from "node:test";
 
 import {
+    getLintFixableBadgeLabel,
     isLintLevel,
     isLintLevelFilter,
     LINT_LEVEL_LABELS,
@@ -106,4 +107,12 @@ void test("parseLintLevelFilter round-trips valid input and nulls invalid input"
     assert.equal(parseLintLevelFilter("anything-else"), null);
     assert.equal(parseLintLevelFilter(""), null);
     assert.equal(parseLintLevelFilter(undefined), null);
+});
+
+void test("getLintFixableBadgeLabel returns null for non-fixable rules", () => {
+    assert.equal(getLintFixableBadgeLabel(null), null);
+});
+
+void test("getLintFixableBadgeLabel returns the canonical fixable label", () => {
+    assert.equal(getLintFixableBadgeLabel("code"), "fixable");
 });
